[Refactor] Extract newEndTime and addSecondsToStar utility functions; update SolCycle component to use new implementations
